requirements.txt
pyautogui==0.9.54
pynput==1.7.6
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.1
openai==1.12.0
